move IsEnabled func
