+++ pkg/tools/tracing/grpcoptions.go
@@ -27,12 +27,12 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/proto"
 
-	"github.com/networkservicemesh/sdk/pkg/tools/log"
+	"github.com/networkservicemesh/sdk/pkg/tools/opentelemetry"
 )
 
 // WithTracing - returns array of grpc.ServerOption that should be passed to grpc.Dial to enable opentelemetry tracing
 func WithTracing() []grpc.ServerOption {
-	if log.IsOpentelemetryEnabled() {
+	if opentelemetry.IsEnabled() {
 		interceptor := func(
 			ctx context.Context,
 			req interface{},
@@ -55,7 +55,7 @@ func WithTracing() []grpc.ServerOption {
 
 // WithTracingDial returns array of grpc.DialOption that should be passed to grpc.Dial to enable opentelemetry tracing
 func WithTracingDial() []grpc.DialOption {
-	if log.IsOpentelemetryEnabled() {
+	if opentelemetry.IsEnabled() {
 		interceptor := func(
 			ctx context.Context,
 			method string,
