+++ pkg/tools/zipkin/exporter.go
@@ -26,11 +26,12 @@ import (
 	"go.opentelemetry.io/otel/sdk/trace"
 
 	"github.com/networkservicemesh/sdk/pkg/tools/log"
+	"github.com/networkservicemesh/sdk/pkg/tools/opentelemetry"
 )
 
 // InitExporter -  returns an instance of Zipkin Exporter.
 func InitExporter(ctx context.Context, exporterURL string) trace.SpanExporter {
-	if !log.IsOpentelemetryEnabled() {
+	if !opentelemetry.IsEnabled() {
 		return nil
 	}
